move jq package into pkg/
